+++ saleor/plugins/avatax/plugin.py
@@ -16,9 +16,6 @@ from prices import Money, TaxedMoney, TaxedMoneyRange
 from ...checkout import base_calculations
 from ...checkout.fetch import fetch_checkout_lines
 from ...checkout.utils import (
-    is_shipping_required as is_shipping_required_for_checkout,
-)
-from ...checkout.utils import (
     log_address_if_validation_skipped_for_checkout,
 )
 from ...core.prices import MAXIMUM_PRICE
@@ -30,7 +27,6 @@ from ...core.taxes import (
 )
 from ...order import base_calculations as order_base_calculation
 from ...order.interface import OrderTaxedPricesData
-from ...order.utils import is_shipping_required as is_shipping_required_for_order
 from ...product.models import ProductType
 from ...tax import TaxCalculationStrategy
 from ...tax.utils import (
@@ -768,10 +764,7 @@ class AvataxPlugin(BasePlugin):
             )
             return None
 
-        is_shipping_required = is_shipping_required_for_checkout(lines_info)
-        if tax_error := self.validate_tax_data(
-            response, lines_info, is_shipping_required
-        ):
+        if tax_error := self.validate_tax_data(response, lines_info):
             self._set_checkout_tax_error(checkout_info, lines_info, tax_error)
             return None
 
@@ -805,8 +798,7 @@ class AvataxPlugin(BasePlugin):
             return None
 
         lines = order.lines.all()
-        is_shipping_required = is_shipping_required_for_order(lines)
-        if tax_error := self.validate_tax_data(response, lines, is_shipping_required):
+        if tax_error := self.validate_tax_data(response, lines):
             self._set_order_tax_error(order, tax_error)
             return None
 
@@ -944,19 +936,12 @@ class AvataxPlugin(BasePlugin):
             cls.validate_authentication(plugin_configuration)
 
     @classmethod
-    def validate_tax_data(
-        cls, tax_data: dict[str, Any], lines: Iterable, is_shipping_required: bool
-    ) -> str:
+    def validate_tax_data(cls, tax_data: dict[str, Any], lines: Iterable) -> str:
         if not tax_data:
             return TaxDataErrorMessage.EMPTY
 
         if cls.check_negative_values_in_plugin_tax_data(tax_data):
             return TaxDataErrorMessage.NEGATIVE_VALUE
-
-        if cls.check_line_number_in_plugin_tax_data(
-            tax_data, lines, is_shipping_required
-        ):
-            return TaxDataErrorMessage.LINE_NUMBER
 
         if cls.check_overflows_in_plugin_tax_data(tax_data):
             return TaxDataErrorMessage.OVERFLOW
@@ -972,22 +957,6 @@ class AvataxPlugin(BasePlugin):
         for line in tax_data.get("lines", []):
             if line.get("lineAmount", 0) < 0:
                 return True
-
-        return False
-
-    @classmethod
-    def check_line_number_in_plugin_tax_data(
-        cls, tax_data: dict[str, Any], lines: Iterable, is_shipping_required: bool
-    ) -> bool:
-        """Check if tax data contains same line number as input data."""
-        if not tax_data:
-            return False
-
-        tax_lines = tax_data.get("lines", [])
-        # shipping data is represented as additional order line
-        expected_lines_length = len(list(lines)) + (1 if is_shipping_required else 0)
-        if len(tax_lines) != expected_lines_length:
-            return True
 
         return False